# config/schema/admin-openapi.yaml
openapi: 3.1.0
info:
  title: AegisProxy administration API
  version: 1.0.0
  description: >-
    Current v1 contract. This document is not served by the daemon. Unless a route explicitly
    says otherwise, the current implementation is available only through the private Unix socket.
    Typed Proxy Host validation and preview are non-persistent.
servers:
  - url: http://localhost
    description: HTTP over the configured private Unix socket
paths:
...
  /v1/web/setup-token:
    post:
      operationId: createWebSetupToken
      description: Authenticated Admin Unix peer only; bearer authentication is rejected.
      x-aegis-authentication: unix-peer-only
      security:
        - unixPeer: []
      responses:
        "201":
          description: Ten-minute one-use plaintext setup token returned once
          headers:
            Cache-Control: {$ref: "#/components/headers/NoStore"}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/IssuedWebSetupToken"}
        default: {$ref: "#/components/responses/Error"}
...
  /v1/certificates/{id}/renew:
    post:
      operationId: renewCertificate
      parameters:
        - {$ref: "#/components/parameters/ObjectId"}
        - {$ref: "#/components/parameters/IfMatch"}
      responses:
        "200": {description: Renewal request persisted}
        default: {$ref: "#/components/responses/Error"}
...
  /v1/tokens:
    get:
      operationId: listTokens
      responses:
        "200":
          description: Hash-free token metadata
          content:
            application/json:
              schema:
                type: array
                maxItems: 1024
                items: {$ref: "#/components/schemas/TokenMetadata"}
        default: {$ref: "#/components/responses/Error"}
    post:
      operationId: createToken
      parameters: [{$ref: "#/components/parameters/IfMatch"}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [user_ref, scopes, expires_unix_secs]
              properties:
                user_ref: {$ref: "#/components/schemas/ObjectId"}
                scopes:
                  type: array
                  minItems: 1
                  maxItems: 53
                  uniqueItems: true
                  items:
                    $ref: "#/components/schemas/TokenScope"
                expires_unix_secs: {type: integer, minimum: 1}
      responses:
        "201":
          description: Plaintext token returned once
          content:
            application/json:
              schema: {$ref: "#/components/schemas/IssuedToken"}
        default: {$ref: "#/components/responses/Error"}
  /v1/tokens/{id}/revoke:
    post:
      operationId: revokeToken
      parameters:
        - {$ref: "#/components/parameters/ObjectId"}
        - {$ref: "#/components/parameters/IfMatch"}
      responses:
        "200": {description: Token revoked}
        default: {$ref: "#/components/responses/Error"}
...
  /v1/backups:
    post:
      operationId: createBackup
      parameters: [{$ref: "#/components/parameters/IfMatch"}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [output]
              properties:
                output: {type: string, maxLength: 4096}
      responses:
        "200": {description: Encrypted backup created}
        default: {$ref: "#/components/responses/Error"}
  /v1/restore/validate:
    post:
      operationId: validateRestore
      parameters: [{$ref: "#/components/parameters/IfMatch"}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [input, identity]
              properties:
                input: {type: string, maxLength: 4096}
                identity: {type: string, description: env:// or absolute file:// secret reference}
      responses:
        "200": {description: Archive authenticated and validated without extraction}
        default: {$ref: "#/components/responses/Error"}
security:
  - bearerToken: []
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: Optional on the private socket; local peer credentials remain mandatory.
    unixPeer:
      type: http
      scheme: aegis-unix-peer
      description: Operating-system peer credentials from the private Unix socket; no HTTP header is accepted.
    browserSession:
      type: apiKey
      in: cookie
      name: __Host-aegis-session
      description: Loopback-browser-only server-side session; bearer authentication is rejected.
  parameters:
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: {type: string, pattern: '^"[0-9]{20}-[0-9a-f]{64}"$'}
...
    ObjectId:
      name: id
      in: path
      required: true
      schema: {type: string, maxLength: 128}
...
  headers:
    NoStore:
      description: Response must not be stored by any cache.
      schema: {type: string, const: no-store}
...
  responses:
    Error:
      description: Stable redacted error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ObjectId:
      type: string
      pattern: '^[a-z][a-z0-9_-]{0,63}$'
...
    TokenScope:
      type: string
      enum: [read_status, read_config, validate_config, preview_config, create_candidate, activate_config, rollback_config, read_revisions, read_proxy_hosts, create_proxy_host, update_proxy_host, delete_proxy_host, activate_proxy_host, rollback_proxy_host, activate_typed_candidate, rollback_typed_revision, read_access_policies, create_access_policy, update_access_policy, delete_access_policy, read_routes, read_upstreams, drain, read_certificates, read_certificate_objects, create_certificate, update_certificate, delete_certificate, read_stream_hosts, create_stream_host, update_stream_host, delete_stream_host, read_discovery_sources, create_discovery_source, update_discovery_source, delete_discovery_source, read_credentials, create_credential, update_credential, revoke_credential, renew_certificate, read_audit, create_backup, validate_restore, manage_identities, read_tokens, create_token, revoke_token, read_users, create_user, update_user, read_roles, create_web_setup_token]
    TokenMetadata:
      type: object
      additionalProperties: false
      required: [id, role, owner_id, user_ref, scopes, expires_unix_secs, revoked]
      properties:
        id: {type: string, pattern: '^[A-Za-z0-9_-]{16}$'}
        role: {type: string, enum: [viewer, auditor, operator, admin]}
        owner_id:
          oneOf:
            - {$ref: "#/components/schemas/ObjectId"}
            - {type: "null"}
        user_ref:
          oneOf:
            - {$ref: "#/components/schemas/ObjectId"}
            - {type: "null"}
        scopes:
          type: array
          maxItems: 53
          uniqueItems: true
          items: {$ref: "#/components/schemas/TokenScope"}
        expires_unix_secs: {type: integer, minimum: 1}
        revoked: {type: boolean}
    IssuedToken:
      type: object
      additionalProperties: false
      required: [token, metadata]
      properties:
        token: {type: string, description: Plaintext returned once and never persisted}
        metadata: {$ref: "#/components/schemas/TokenMetadata"}
...
    IssuedWebSetupToken:
      type: object
      additionalProperties: false
      required: [token, owner_id, expires_unix_secs]
      properties:
        token:
          type: string
          pattern: '^[A-Za-z0-9_-]{43}$'
          description: Plaintext returned once, retained only as an in-memory SHA-256 digest
        owner_id: {$ref: "#/components/schemas/ObjectId"}
        expires_unix_secs: {type: integer, minimum: 1}
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, details, request_id]
          properties:
            code: {type: string}
            message: {type: string}
            details: {type: array, items: {type: object}}
            request_id: {type: string}